Add compute tag for scalar source coupled to curvature in ScalarTensor
